Replace MercuryMover with Hammerspoon
I have for a long time used an app called MercuryMover to resize and
move windows around.  Unfortunately, it stopped working under macOS
Sierra. Other apps, such as Spectacle and Moom, have a very limited set
of window sizes that just didn't work for my (overly specific) needs.

Hammerspoon is an automation tool. I don't (yet?) use it for much, but
one thing it can do is resize windows, and I have replicated the core
functionality of MercuryMover that I miss in it.